refactor(cache): move redis to cache package and add main function
