+++ pkg/cache/cache_impl.go
@@ -1,4 +1,4 @@
-package redis
+package cache
 
 import (
 	"context"
